Review Request 8332 add docker configurations

+++ le-dev/scripts/stop-cluster.sh
@@ -1,6 +1,11 @@
+#!/bin/bash
+
 # Test for required env variables
 printf "%s\n" "${HADOOP_HOME:?You must set HADOOP_HOME}"
-printf "%s\n" "${ZOOKEEPER_HOME:?You must set ZOOKEEPER_HOME}"
+
+if [ "${LE_USING_DOCKER}" != "true" ]; then
+    printf "%s\n" "${ZOOKEEPER_HOME:?You must set ZOOKEEPER_HOME}"
+fi
 
 $HADOOP_HOME/sbin/hadoop-daemon.sh stop namenode
 $HADOOP_HOME/sbin/hadoop-daemon.sh stop datanode
@@ -9,10 +14,17 @@ $HADOOP_HOME/sbin/yarn-daemon.sh stop nodemanager
 $HADOOP_HOME/sbin/yarn-daemon.sh stop timelineserver
 $HADOOP_HOME/sbin/mr-jobhistory-daemon.sh stop historyserver
 $HADOOP_HOME/sbin/kms.sh stop
-$ZOOKEEPER_HOME/bin/zkServer.sh stop 
 
-DYNAMO_PID=`jps | grep DynamoDBLocal.jar | awk '{print $1}'`
-if [ -n "$DYNAMO_PID" ]; then
-    echo "Killing Dynamo with pid $DYNAMO_PID"
-    kill $DYNAMO_PID 
+if [ "${LE_USING_DOCKER}" = "true" ]; then
+    echo "You are in Docker environment, please use dk-stop to stop docker compose pod"
+else
+    $ZOOKEEPER_HOME/bin/zkServer.sh stop
+    DYNAMO_PID=`jps | grep DynamoDBLocal.jar | awk '{print $1}'`
+    if [ -n "$DYNAMO_PID" ]; then
+        echo "Killing Dynamo with pid $DYNAMO_PID"
+        kill $DYNAMO_PID
+    fi
 fi
+
+
+
